+++ src/models/carousel.ts
@@ -1,0 +1,3 @@
+export interface CarouselProps {
+    slides: string[];
+}
